Safari-611.1.10/LayoutTests/fast/url/user-visible/srb-expected.txt

Test IDN behavior for .срб top level domain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test('http://православнајапородица.срб/') is 'http://православнајапородица.срб/'
PASS test('http://www.свеовде.од.срб/') is 'http://www.свеовде.од.срб/'
PASS successfullyParsed is true

TEST COMPLETE
